+++ src/main/kotlin/icu/windea/pls/lang/refactoring/rename/naming/AutomaticDefinitionGeneratedModifiersRenamer.kt
@@ -16,11 +16,11 @@ import icu.windea.pls.script.psi.ParadoxDefinitionElement
  */
 class AutomaticDefinitionGeneratedModifiersRenamer(element: PsiElement, newName: String) : AutomaticRenamer() {
     init {
-        element as ParadoxDefinitionElement
-        val allRenames = mutableMapOf<PsiElement, String>()
+        val allRenames = mutableMapOf<PsiNamedElement, String>()
         prepareRenaming(element, newName, allRenames)
         for ((key, value) in allRenames) {
-            myElements.add(key as PsiNamedElement)
+            ProgressManager.checkCanceled()
+            myElements += key
             suggestAllNames(key.name, value)
         }
     }
@@ -35,7 +35,8 @@ class AutomaticDefinitionGeneratedModifiersRenamer(element: PsiElement, newName:
 
     override fun entityName() = PlsBundle.message("rename.definition.generatedModifiers.entityName")
 
-    private fun prepareRenaming(element: ParadoxDefinitionElement, newName: String, allRenames: MutableMap<PsiElement, String>) {
+    private fun prepareRenaming(element: PsiElement, newName: String, allRenames: MutableMap<PsiNamedElement, String>) {
+        if (element !is ParadoxDefinitionElement) return
         val definitionInfo = element.definitionInfo ?: return
         val infos = definitionInfo.modifiers.orNull() ?: return
         for (info in infos) {