+++ src/main/kotlin/icu/windea/pls/lang/refactoring/rename/naming/AutomaticDynamicValueRelatedLocalisationsRenamer.kt
@@ -6,12 +6,15 @@ import com.intellij.psi.PsiNamedElement
 import com.intellij.refactoring.rename.naming.AutomaticRenamer
 import icu.windea.pls.PlsBundle
 import icu.windea.pls.core.orNull
-import icu.windea.pls.lang.search.ParadoxScriptedVariableSearch
+import icu.windea.pls.lang.psi.mock.ParadoxDynamicValueElement
+import icu.windea.pls.lang.search.ParadoxLocalisationSearch
 import icu.windea.pls.lang.search.selector.contextSensitive
 import icu.windea.pls.lang.search.selector.selector
-import icu.windea.pls.script.psi.ParadoxScriptScriptedVariable
 
-class AutomaticScriptedVariablesRenamer(element: PsiElement, newName: String) : AutomaticRenamer() {
+/**
+ * 用于在重命名动态值时，自动重命名相关本地化（如果存在且需要）。
+ */
+class AutomaticDynamicValueRelatedLocalisationsRenamer(element: PsiElement, newName: String) : AutomaticRenamer() {
     init {
         val allRenames = mutableMapOf<PsiNamedElement, String>()
         prepareRenaming(element, newName, allRenames)
@@ -26,23 +29,20 @@ class AutomaticScriptedVariablesRenamer(element: PsiElement, newName: String) : 
 
     override fun allowChangeSuggestedName() = false
 
-    override fun getDialogTitle() = PlsBundle.message("rename.scriptedVariable.overrides.title")
+    override fun getDialogTitle() = PlsBundle.message("rename.dynamicValue.relatedLocalisations.title")
 
-    override fun getDialogDescription() = PlsBundle.message("rename.scriptedVariable.overrides.desc")
+    override fun getDialogDescription() = PlsBundle.message("rename.dynamicValue.relatedLocalisations.desc")
 
-    override fun entityName() = PlsBundle.message("rename.scriptedVariable.overrides.entityName")
+    override fun entityName() = PlsBundle.message("rename.dynamicValue.relatedLocalisations.entityName")
 
     private fun prepareRenaming(element: PsiElement, newName: String, allRenames: MutableMap<PsiNamedElement, String>) {
-        if (element !is ParadoxScriptScriptedVariable) return
-        val name = element.name?.orNull() ?: return
+        if (element !is ParadoxDynamicValueElement) return
+        val name = element.name.orNull() ?: return
         ProgressManager.checkCanceled()
-        val selector = selector(element.project, element).scriptedVariable().contextSensitive()
-        val targets = mutableSetOf<ParadoxScriptScriptedVariable>()
-        ParadoxScriptedVariableSearch.searchLocal(name, selector).findAll().let { targets.addAll(it) }
-        ParadoxScriptedVariableSearch.searchGlobal(name, selector).findAll().let { targets.addAll(it) }
+        val selector = selector(element.project, element).localisation().contextSensitive()
+        val targets = ParadoxLocalisationSearch.searchNormal(name, selector).findAll()
         for (target in targets) {
             ProgressManager.checkCanceled()
-            if (target == element) continue
             allRenames[target] = newName
         }
     }