+++ src/main/kotlin/icu/windea/pls/lang/refactoring/rename/naming/AutomaticScriptedVariableRelatedLocalisationsRenamerFactory.kt
@@ -11,6 +11,9 @@ import icu.windea.pls.lang.util.ParadoxLocaleManager
 import icu.windea.pls.lang.util.ParadoxScriptedVariableManager
 import icu.windea.pls.script.psi.ParadoxScriptScriptedVariable
 
+/**
+ * 用于在重命名封装变量时，自动重命名相关本地化（如果存在且需要）。
+ */
 class AutomaticScriptedVariableRelatedLocalisationsRenamerFactory : AutomaticRenamerFactory {
     override fun isApplicable(element: PsiElement): Boolean {
         if (element !is ParadoxScriptScriptedVariable) return false