+++ src/main/kotlin/icu/windea/pls/lang/refactoring/rename/naming/AutomaticScriptedVariablesRenamerFactory.kt
@@ -17,7 +17,6 @@ class AutomaticScriptedVariablesRenamerFactory : AutomaticRenamerFactory {
     override fun isApplicable(element: PsiElement): Boolean {
         if (element !is ParadoxScriptScriptedVariable) return false
         val name = element.name?.orNull() ?: return false
-
         val selector = selector(element.project, element).scriptedVariable()
         val processor = Processors.duplicate<ParadoxScriptScriptedVariable>()
         ParadoxScriptedVariableSearch.searchLocal(name, selector).process(processor)