+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 - [x] 修复：修复条件块中的参数在调用时是否必须传入的判断逻辑的相关问题 / FIX: Fixed a bug that the parameters in the condition block were not correctly identified #272
 - [x] 优化：格式化 CWT 文件时，`!=` `<>` 周围始终保留空格，避免格式化前后的语法不等效 / OP: Optimize the formatting of CWT files, always keep spaces around `!=` and `<>`, to avoid syntax inequality before and after formatting
 - [x] 优化：与重命名有关的相关优化与BUG修复 / OP: Optimizations and bug fixes related to renaming
-- [x] 优化：补充更多缺失的 AtomaticRenamer（关联重命名） / OP: Add more missing `AutomaticRenamer` (association-rename)
+- [x] 优化：补充更多缺失的 `AutomaticRenamer`（关联重命名） / OP: Add more missing `AutomaticRenamer` (association-rename)
 - [x] 优化：为定义注入检测子类型 / OP: Detect subtypes for definition injection #274
 - [x] 优化：将模式为 `REPLACE_OR_CREATE` 的定义注入识别为定义声明 / OP: Detect definition injections with mode `REPLACE_OR_CREATE` as definition declarations #273
 - [x] 新功能：新增 `PutMembersOnOneLineIntention` 和 `PutMembersOnSeparateLinesIntention` / NEW: add `PutMembersOnOneLineIntention` and `PutMembersOnSeparateLinesIntention`